+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - fix(translation.fr): Correct small French translation issues (#87)
 - fix(pagination): Fix database integer overflow with pagination (#93)
 - fix(session): Fix missing `SameSite` attribute in session cookie (#89)
+- fix(headers): Fix secutiry issues by adding secutiry headers (#90, #91)
 
 ### Changed